refactor(core): move component types and Phase from basalt-ecs to basalt-core
Component types (Position, Velocity, Inventory, etc.), the Component
marker trait, EntityId, and Phase are domain types — they belong in
basalt-core, not in the storage engine. Split into focused files:
components/spatial.rs, identity.rs, item.rs, inventory.rs.

basalt-ecs re-exports from basalt-core for backward compatibility.
