+++ crates/basalt-ecs/src/lib.rs
@@ -1,4 +1,4 @@
-//! In-house Entity Component System for the Basalt game loop.
+//! Generic Entity Component System — pure storage engine.
 //!
 //! Intentionally simple: `HashMap` per component type, not archetype-based.
 //! Sufficient for thousands of entities without the complexity of a full
@@ -7,14 +7,12 @@
 //! - **Entity**: a unique [`EntityId`] (u32). Just an ID, no data.
 //! - **Component**: a typed struct stored in a `HashMap<EntityId, T>`.
 //! - **System**: a function that runs each tick with access to the ECS.
+//!
+//! This crate has **zero domain knowledge** — no Minecraft types, no
+//! game-specific components. Component types are defined in `basalt-core`.
 
-mod components;
 mod ecs;
 mod system;
 
-pub use components::{
-    BoundingBox, DroppedItem, EntityKind, Health, Inventory, Lifetime, OpenContainer, PickupDelay,
-    PlayerRef, Position, Rotation, Velocity,
-};
 pub use ecs::{Component, Ecs, EntityId};
 pub use system::{Phase, SystemAccess, SystemBuilder, SystemDescriptor, SystemRunner};
